Split tests about DeltaMergeStore
