+++ dbms/src/Storages/DeltaMerge/CMakeLists.txt
@@ -20,26 +20,24 @@ add_subdirectory(./Remote/Proto)
 
 add_headers_and_sources(delta_merge .)
 add_headers_and_sources(delta_merge ./BitmapFilter)
+add_headers_and_sources(delta_merge ./ColumnFile)
 add_headers_and_sources(delta_merge ./Index)
-add_headers_and_sources(delta_merge ./Index/VectorIndex)
-add_headers_and_sources(delta_merge ./Index/VectorIndex/Stream)
-add_headers_and_sources(delta_merge ./Index/InvertedIndex)
+add_headers_and_sources(delta_merge ./Decode)
+add_headers_and_sources(delta_merge ./Delta)
+add_headers_and_sources(delta_merge ./DeltaIndex)
+add_headers_and_sources(delta_merge ./File)
 add_headers_and_sources(delta_merge ./Filter)
 add_headers_and_sources(delta_merge ./FilterParser)
-add_headers_and_sources(delta_merge ./File)
-add_headers_and_sources(delta_merge ./ColumnFile)
-add_headers_and_sources(delta_merge ./Delta)
+add_headers_and_sources(delta_merge ./Index/InvertedIndex)
+add_headers_and_sources(delta_merge ./Index/VectorIndex)
+add_headers_and_sources(delta_merge ./Index/VectorIndex/Stream)
 add_headers_and_sources(delta_merge ./ReadThread)
 add_headers_and_sources(delta_merge ./Remote)
 add_headers_and_sources(delta_merge ./Remote/DataStore)
-add_headers_and_sources(delta_merge ./Decode)
 add_headers_and_sources(delta_merge ./StoragePool)
+add_headers_and_sources(delta_merge ./VersionChain)
 
 add_library(delta_merge ${delta_merge_headers} ${delta_merge_sources})
 target_link_libraries(delta_merge PRIVATE dbms page)
 
 add_subdirectory(workload)
-
-if (ENABLE_TESTS)
-    add_subdirectory(tests)
-endif ()
