+++ dbms/src/Storages/DeltaMerge/DeltaIndex/DeltaIndexManager.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include <Storages/DeltaMerge/DeltaIndex.h>
+#include <Storages/DeltaMerge/DeltaIndex/DeltaIndex.h>
 #include <common/logger_useful.h>
 
 namespace DB
