+++ dbms/src/Storages/DeltaMerge/DeltaIndex/tests/gtest_dm_delta_tree.cpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <Storages/DeltaMerge/DeltaIndex/DeltaTree.h>
 #include <Storages/DeltaMerge/DeltaMergeDefines.h>
-#include <Storages/DeltaMerge/DeltaTree.h>
 #include <TestUtils/TiFlashTestBasic.h>
 
 namespace DB::DM::tests
