+++ dbms/src/Storages/DeltaMerge/Index/VectorIndex/tests/bench_vector_index.cpp
