+++ dbms/src/Storages/DeltaMerge/tests/gtest_dm_utils.cpp
@@ -17,20 +17,18 @@
 #include <Storages/DeltaMerge/tests/DMTestEnv.h>
 #include <TestUtils/TiFlashTestBasic.h>
 
-namespace DB
+namespace DB::DM::tests
 {
-namespace DM
-{
-namespace tests
-{
+
 TEST(RowKeyFilterTest, FilterSortedBlock)
 {
     const size_t num_rows_write = 100;
-    RowKeyRanges ranges;
     // create ranges unsorted and overlapped with each other
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(20, 50)));
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(0, 30)));
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(70, 90)));
+    RowKeyRanges ranges{
+        RowKeyRange::fromHandleRange(HandleRange(20, 50)),
+        RowKeyRange::fromHandleRange(HandleRange(0, 30)),
+        RowKeyRange::fromHandleRange(HandleRange(70, 90)),
+    };
     Block block = DMTestEnv::prepareSimpleWriteBlock(0, num_rows_write, false);
     auto filtered_block = RowKeyFilter::filterSorted(ranges, std::move(block), 0);
     ASSERT_EQ(filtered_block.rows(), 70);
@@ -39,11 +37,12 @@ TEST(RowKeyFilterTest, FilterSortedBlock)
 TEST(RowKeyFilterTest, FilterUnsortedBlock)
 {
     const size_t num_rows_write = 100;
-    RowKeyRanges ranges;
     // create ranges unsorted and overlapped with each other
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(20, 50)));
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(0, 30)));
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(70, 90)));
+    RowKeyRanges ranges{
+        RowKeyRange::fromHandleRange(HandleRange(20, 50)),
+        RowKeyRange::fromHandleRange(HandleRange(0, 30)),
+        RowKeyRange::fromHandleRange(HandleRange(70, 90)),
+    };
     Block block = DMTestEnv::prepareSimpleWriteBlock(0, num_rows_write, false);
     auto filtered_block = RowKeyFilter::filterUnsorted(ranges, std::move(block), 0);
     ASSERT_EQ(filtered_block.rows(), 70);
@@ -52,11 +51,12 @@ TEST(RowKeyFilterTest, FilterUnsortedBlock)
 TEST(RowKeyFilterTest, FilterSortedBlockCommonHandle)
 {
     const size_t num_rows_write = 100;
-    RowKeyRanges ranges;
     // create ranges unsorted and overlapped with each other
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(20, 50), true));
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(0, 30), true));
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(70, 90), true));
+    RowKeyRanges ranges{
+        RowKeyRange::fromHandleRange(HandleRange(20, 50), true),
+        RowKeyRange::fromHandleRange(HandleRange(0, 30), true),
+        RowKeyRange::fromHandleRange(HandleRange(70, 90), true),
+    };
     Block block = DMTestEnv::prepareSimpleWriteBlock(
         0,
         num_rows_write,
@@ -74,11 +74,12 @@ TEST(RowKeyFilterTest, FilterSortedBlockCommonHandle)
 TEST(RowKeyFilterTest, FilterUnsortedBlockCommonHandle)
 {
     const size_t num_rows_write = 100;
-    RowKeyRanges ranges;
     // create ranges unsorted and overlapped with each other
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(20, 50), true));
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(0, 30), true));
-    ranges.emplace_back(RowKeyRange::fromHandleRange(HandleRange(70, 90), true));
+    RowKeyRanges ranges{
+        RowKeyRange::fromHandleRange(HandleRange(20, 50), true),
+        RowKeyRange::fromHandleRange(HandleRange(0, 30), true),
+        RowKeyRange::fromHandleRange(HandleRange(70, 90), true),
+    };
     Block block = DMTestEnv::prepareSimpleWriteBlock(
         0,
         num_rows_write,
@@ -93,6 +94,4 @@ TEST(RowKeyFilterTest, FilterUnsortedBlockCommonHandle)
     ASSERT_EQ(filtered_block.rows(), 70);
 }
 
-} // namespace tests
-} // namespace DM
-} // namespace DB
+} // namespace DB::DM::tests
